feat: bruddlize create req view + remove unused code
