+++ src/app/(mobile-ui)/request/create/page.tsx
@@ -1,5 +1,6 @@
 import { generateMetadata } from '@/app/metadata'
-import { CreateRequestLink } from '@/components/Request/Create/Create'
+import PageContainer from '@/components/0_Bruddle/PageContainer'
+import { RequestCreateView } from '@/components/Request/Create/Request.create.view'
 
 export const metadata = generateMetadata({
     title: 'Request Payment | Peanut',
@@ -9,5 +10,9 @@ export const metadata = generateMetadata({
 })
 
 export default function RequestCreate() {
-    return <CreateRequestLink />
+    return (
+        <PageContainer>
+            <RequestCreateView />
+        </PageContainer>
+    )
 }
