--- src/components/Request/Create/Views/index.ts
@@ -1,2 +1,0 @@
-export * from './Initial.view'
-export * from './Success.view'
