--- src/components/Request/Create/index.ts
@@ -1,1 +1,0 @@
-export * from './Create'
