+++ src/components/Request/index.ts
@@ -1,2 +1,1 @@
 export * from './Pay'
-export * from './Create'
